Organized components and left fixlist for groupmembers
There is a readme.fixlist that describes various problems I was unable to fix due to lack of HTML knowlege. I did leave instructions on what the issue was when I was able to discover it. Also, the structure of site was changed but a lot of the navigation and code in other files wasn't fixed to reflect this.  It needs to either be returned back to the old style (which means creating the necessary coimponents for those pages too) or the new changes need to be reflected in the consturction of the nav bars and the href's inside certain components. Again, no idea how the new style was pulled off, so left notes and pointers.

This is in addition to the many things I was able to fix. Please leave file names as is. Please name new files similarly if possible, its cleaner and it reflects the conventions of react and web development better.